requirements.txt
streamlit
google-generativeai
python-dotenv
python-docx
pandas
matplotlib